move all the circuit breaker code
